requirements.txt
streamlit
paddleocr
paddlepaddle
numpy
Pillow
